fix: :bug: export slash print
